+++ addOns/zest/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Allow other add-ons to create a Zest script from a list of messages.
+
 ### Changed
 - Use Semantic Version.
 - Maintenance changes.